- [CLN] Pykeops evaluator
- [CLN] Pykeops code

+++ gempy_engine/config.py
@@ -9,7 +9,7 @@ class AvailableBackends(Enum):
 
 DEBUG_MODE = True
 DEFAULT_BACKEND = AvailableBackends.numpy
-DEFAULT_DTYPE = "float64"
+DEFAULT_DTYPE = "float32"
 
 is_numpy_installed = find_spec("numpy") is not None
 is_tensorflow_installed = find_spec("tensorflow") is not None
